AntlrLimeModelBuilder: use new annotations for description of property
First lines of a comment are used to describe a
property (for Swift and Dart) as well as to describe
setter's parameter name and return type of getter
(in Java and C++). This is confusing.

This change is the beginning of series that splits
the description into 'value' and 'description' annotations.

This change introduces usage of 'value' annotation for
the name and meaning of the given property. Moreover, it
introduces ingestion of 'description' annotation.

The 'value' annotation is used as follows:
 a) to add comment when defining property in Swift/Dart
 b) to describe setter's parameter and getter return type
    in Java/C++

The 'description' annotation is used as follows:
 a) to add comment when defining property in Swift/Dart
 b) to add comment when defining getter/sett in Java/C++

This change adjusts also the input of smoke tests to use
the new annotations.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>