smoke-tests: adjust output for documenting properties
In the previous commit the usage of 'description' annotation
was added to the input LIME files. However, the new field is
not yet used by mustache templates. Therefore, some information
has been stripped from the output files.

The good thing is that the information was removed from the
description of parameters and return types of getters/setters
for Java/C++.

The bad thing is that the information was removed from the
declaration comment of properties in Swift/Dart.

In next patches mustache files will be adjusted to obtain the
correct behavior.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>