Dart: 'PositionalDefaults' annotation and null collection yields incorrect code
This change is used to showcase the invalid behavior of
Dart generator. Usage of 'PositionalDefaults' annotation
combined with nullable collection field that has 'null'
default value yields the code that does not compile.

The generator blindly applies 'const' keyword before
the default value for collections. This is done in
'DartStructConstructors.mustache' depending on the
field LIME type. However, a special 'null' value is
not taken into account. It seems to be invalid.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>