Dart: 'Immutable' annotation and collections yield incorrect code
This change is used to showcase the invalid behavior of
Dart generator. Usage of 'Immutable' annotation combined
with collection fields that have default values yields the
code that does not compile.

The generator does not apply 'const' keyword before the calls
to constructors of collections. It is required, because the
constructor of 'Immutable' structure is marked as 'const'.

The same happens in an explicit field constructor is used.
Because of 'Immutable' usage it is marked as a const, but
the default values are not initialized with values prepended
by 'const'.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>