Dart: fix problems with default values that use collection types (#1622)
This series fixes problems that occur when either @Immutable
or @PositionalDefaults annotation is used in combination with
collection types.

-------- @PositionalDefaults --------
The usage of 'PositionalDefaults' annotation
combined with nullable collection field that has 'null'
default value yields the code that does not compile.

The generator blindly applies 'const' keyword before
the default value for collections. A special 'null' value is
not taken into account.

-------- @Immutable --------
Usage of 'Immutable' annotation combined
with collection fields that have default values yields the
code that does not compile.

The generator does not apply 'const' keyword before the calls
to constructors of collections. It is required, because the
constructor of 'Immutable' structure is marked as 'const'.

The same happens when an explicit field constructor is used.
Because of 'Immutable' usage it is marked as a const, but
the default values are not initialized with values prepended
by 'const'.

-------- Content of change --------

- New smoke and functional tests that were used to confirm
   the invalid behavior.
- Adjustments of 'DartStructConstructors.mustache' file to
   correctly prepend 'const' keyword when needed and to treat
   'null' as it should be treated.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>