Dart: 'Immutable' annotation and immutable struct fields do not compile
This change is used to showcase the invalid behavior of
Dart generator. Usage of 'Immutable' annotation combined
with field, which has default value and uses immutable
struct type yields the code that does not compile.

The generator does not apply 'const' keyword before the calls
to constructors of structures. It is required, because the
constructor of 'Immutable' structure is marked as 'const'.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>