LimeFunctionsValidatorDocsTest: use more realistic paths to elements
This change adjusts the unit tests to use more
realistic paths to elements from LIME model.

The adjusted paths are better, because the validator
uses 'needsParametersDocumentation()' to check if the
validation of docs needs to be run.

The adjusted paths correspond to the ones that would
be built by AntlrLimeModelBuilder.kt.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>