Gluecodium.kt: validate functions just once
'LimeFunctionsValidator' uses 'typeRef' member while performing validation.
Therefore, it shall be present only in 'getTypeRefDependentValidators()'.

However, it is present also in 'getIndependentValidators()' which means,
that Gluecodium validates functions twice.

The investigation of GIT history revealed that addition of functions
validator to 'independent validators' list was done accidentally.

This change ensures, that the mentioned validator is run only once.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>