Confirm mismatch between generated default values
This change is used to showcase the invalid behavior
of the code generator.

It implements new functional tests that verify the
default values that are set in LIME file for nullable
fields.

The default value is invalid in the case of default
constructed object in C++. There is a mismatch between
C++ and Java/Swift/Dart.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>