Showcase invalid constructor call in C++ for default value
This change introudces the same LIME structures
that the previous commit, but in smoke tests.

The following LIME syntax is used:
```
  someField1: SomeImmutableStructWithDefaults? = {}
  someField2: ImmutableStructWithCollections? = {}
```

It results in the following invalid C++ code (fields are nullopt):
```
  some_field1
    = std::optional<SomeImmutableStructWithDefaults>{};

  some_field2
    = std::optional<ImmutableStructWithCollections>{};
```

The generated C++ code for default values initializes fields
to nullopt instead of default-constructed structure. This
behavior is invalid and does not occur in other languages.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>