Improve structured comments syntax for properties  (#1617)
The first part of comment used to describe a property (lines before get/set annotations)
was used in a few places:
 - to document the property when declaring it in Swift and Dart
 - to document the parameter used in setter in the case of C++ and Java
 - to document the return value of getter in the case of C++ and Java

In some cases it looked strange. For instance, when the user wanted to leave some additional,
extended information about the property, that was necessary to use it correctly, then it was
appended to the description of the return type and parameters of getters and setters.

Moreover, it was hard for users to remember how these lines are rendered in the case of four
output languages.

In order to ease the documentation of properties, this change introduces two new annotations:
'value' and 'description'.

The structured comments for properties can be formed as follows:

'value':
- purpose: short description of property (ideally single line)
- usage: declaration of property (Swift/Dart) or setter_param/getter_return (Java/C++)

'description':
- purpose: extended information about property that is required to use it correctly
- usage: declaration of property (Swift/Dart), declaration of getter/setter (Java/C++)

'get':
 - purpose: documents getter function
 - usage: declaration of getter function

'set':
 - purpose: documents setter function
 - usage: declaration of setter function

Moreover, the validation was implemented. The rules are as follows:
 - the user cannot start a comment without annotation and then use later 'value'
 - if the user forgets to use 'value' then warning will be generated
 - if the user wants, the warning related to missing 'value' can be treated as error via werror flag
 - the old way of generating comment will be used when the comment does not use 'value' but provides
    information without annotation -- however, the warning/error will be generated as well

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>