Dart: allow arbitrary defaults for 'PositionalDefaults'
This change introduces the fix for the bug showcased
in the previous commit.

In order to be able to generate code, which compiles
when 'PositionalDefaults' annotation is used together
with default values of structure types that do not
provide const constructors, the optional values
have been used.

When Gluecodium encounters such type that does not
provide const constructor it makes it optional in
positional defaults constructor and uses null as
default.

Later, on initializer list it checks if user provided
value different than null and if not, then initializes
the field with the actual default value.

This way the limitation related to const constructors
is surpassed.

However, if the user wants to have nullable type with
default different than null, then the behavior may be
surprising.

Therefore, in next patches the required validation will
be implemented for DartGenerator.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>