Dart: confirm invalid behavior for BLOB type
The constructors of 'Uint8List' type are not const.
This type is used when BLOB is specified in LIME
file.

Therefore, the usage of BLOB type for a field of
struct that is marked as 'PositionalDefaults' and
specifying a default value of this field leads to
compilation error.

This change is used to document the invalid behavior.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>