Dart: make generated LazyList 'Finalizable'
If the generated class is not marked with 'Finalizable',
then when a method is invoked on its object, the object can be
garbage collected in the middle of execution of the method.

The consequences can be dramatic -- the finalizer registered for
the object may be executed before the native function call inside
the member methods. This can lead to segmentation fault.

The 'Finalizable' interface ensures, that the object outlives the
native function call.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>