functional_tests: do not depend on local time zone in Dart tests
In 'date_defaults.dart' we expected numeric values of
date fields that were hardcoded for UTC time zone.

However, when 'DateTime' was created without specified
UTC time zone, then we received numeric values that were
specific for a local time zone of the runner of tests.

This change makes the test stable and removes the
implicit dependency to the local time zone.

Since this commit the test passes even when run on
a machine, which uses different default time zone.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>