smoke tests: add example of class hierarchy with static methods
This change introduces a new LIME file to smoke
tests, which causes generation of the code for
base and derived classes, which have static methods.

The classes are generated only for Dart and the intention
is to show how 'visibleForTesting' annotation is used.

Sadly, the base type annotated as visible for testing is
used as a base class of the derived impl type. It causes
linter warnings.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>