Dart: do not use 'visibleForTesting' class as a base
In order to allow users to mock static functions in classes
the 'prototype' approach was implemented in the past. It uses
prototype class to redirect static function calls.

A class with '$Impl' suffix is created as a prototype. It is
annotated as 'visibleForTesting' and exported from the library.

The user may provide its own implementation in tests and this
way redirect the calls to the mock (after setting it as prototype).

Sadly, when a class is derived from the class, which has static
methods, then its '$Impl' derives from the '$Impl' of base class.
This leads to linter warnings: 'invalid_use_of_visible_for_testing_member'.

This change introduces a new type called '$HiddenImpl', which is
defined only when the class is open and has static functions.
The '$HiddenImpl' is not exported and therefore, the user cannot use it
directly. Still, the user may use the old approach to redirect the calls
to mock.

The '$HiddenImpl' is derived from an ordinary '$Impl' and its usage does
not generate linter warnings when the '$Impl' of derived class uses it.

The '$HiddenImpl' is used only to provide implementation of methods from
the base class.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>