JavaNameRules: remove dead code
The overload of 'getName()' function for 'LimeLambdaParameter'
is not used anywhere.

Initially, it was used by 'JavaModelBuilder', but this class
was removed in the past and it no longer exists.

Currently, the type 'LimeLambdaParameter' is not used in Java
generator and its templates, because 'JavaLambda.mustache'
utilizes 'LimeLambda.asFunction()', which converts the
mentioned type via 'LimeLambdaParameter.asLimeParameter()'.

The indices of lambda parameters are encoded in the path and
stored via 'AntlrLimeModelBuilder.exitLambda()'.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>