Dart: do not generate `default` clause for enumerations (#1641)
----- Motivation -----
In the case of enumerations we match all the possible
matchable values in `case` clauses. Then, we generate
`default` clause, but the linter is able to see, that
this is a dead code.

In such case `unreachable_switch_default` warning is
raised by the linter.

----- Important -----
The above is true if we generate an authentic `enum` in Dart.
In the case when aliasing enumerators are needed, then
we generate an enum-like class. In such case `default` clause
is needed, because we operate on integers and compiler assumes,
that switch-case is not exhaustive.

----- Implemented solution -----
Removed generation of redundant `default` clause when we use
an authentic Dart `enum` to avoid linter warnings.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>