Dart: allow conditionally disabling 'Finalizable' marker usage
It turned out that in certain rare cases
Dart compiler experiences internal compiler
error related to scope building when 'Finalizable'
is used.

The fix for the problems is available in Dart SDK 3.6.0+.
Therefore, to ease the transition phase for the users a
new CLI flag called 'dartdisablefinalizablemarker' was
introduced. It allows users to conditionally disable
the usage of 'Finalizable' marker interface.

It is discouraged to use this flag, because when 'Finalizable'
marker interface is not used a rare race conditions between
garbage collection and native finalizers can be used.

The flag is designed to help during the transition phase to
Dart SDK 3.6.0+.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>