cmake-tests: introduce a new test for DART_DISABLE_FINALIZABLE_MARKER
This change introduces a new cmake tests, which
verifies that 'GLUECODIUM_DART_DISABLE_FINALIZABLE_MARKER'
flag works as expected.

When it is enabled, then generated file for Dart should not
use 'Finalizable' interface marker from Dart FFI.

The positive test is not added, because this flag is disabled
by default -- this logic is verified in smoke tests.

Moreover, this commit introduces a new utility tool for tests.
It allows checking if the generated file contains a given string.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>