JniWrapperCache: delete weak ref when object was garbage collected
When trying to obtain a Java object from cache
we may encounter a situation in which we cannot
create local reference, because the Java object
was garbage collected. In such a case obtained
local reference points to NULL.

We should delete weak reference and the entry from
cache, because if the check is performed in the
'convert_to_jni()' function we create a new Java object
and try to cache it in JniWrapperCache. The function which
adds to cache does not check if it contains any record for
given pointer.

We may encounter the situation in which we override handle
stored in cache without deleting it. This would lead to weak
reference handle leak.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>