Java: extend JniWrapperCache checks to identify double dispose problem
This change extends debug checks in JniWrapperCache class
to show the problem related to double dispose.

When we remove an entry from JniWrapperCache in 'get_cached_wrapper_impl()'
because the reference was invalidated we must anticipate, that the cleanup
will be scheduled for this non-existent object. It may cause problems, because
the cleanup of weak reference entry for the old object may remove the entry
for the new entry.

This commit extends checking and again makes 'stressTestJniWeakReferenceCache()'
test to fail.

In next commit the fix will be implemented to prevent such double-dispose.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>