Java: show leak of handles to weak references
According to Java documentation of 'PhantomReference' class, which is used to
schedule post-mortem cleanup of resources of C++ objects tied with Java objects
and cleanup of the JniWrapperCache entry, there can be a time gap between clearing
PhantomReference of Java object and putting PhantomReference object in the registered
queue.

During this time window the user of JniWrapperCache should not be capable of getting
Java object from weak reference handle stored in wrapper cache map, because JNI weak
reference returns null -- it is treated like PhantomReference and also cleared by GC.

Sadly, the function, which adds to JniWrapperCache does not check if there is an entry for a given
native pointer. As a result, we may override existing weak reference handle -- therefore, we would
never destroy it. This results in a weak reference handle leak.

The situation may happen if we query the same object multiple times. For instance when we have a
C++ class, which has a member of object type that can be obtained via getter function. Calling this
getter multiple times may cause weak reference handles leak.

This commit brings a new functional test, which triggers the leak.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>