Java: remove leak of JNI weak references in JniWrapperCache (#1648)
----- Motivation -----
According to Java documentation of 'PhantomReference' class, which is used to
schedule post-mortem cleanup of resources of C++ objects tied with Java objects
and cleanup of the JniWrapperCache entry, there can be a time gap between clearing
PhantomReference of Java object and putting PhantomReference object in the registered
queue.

During this time window the user of JniWrapperCache should not be capable of getting
Java object from weak reference handle stored in wrapper cache map, because JNI weak
reference returns null -- it is treated like PhantomReference and also cleared by GC.

Sadly, the function, which adds to JniWrapperCache does not check if there is an entry for a given
native pointer. As a result, we may override existing weak reference handle -- therefore, we would
never destroy it. This results in a weak reference handle leak.

The situation may happen if we query the same object multiple times. For instance when we have a
C++ class, which has a member of object type that can be obtained via getter function. Calling this
getter multiple times may cause weak reference handles leak.

----- Implemented changes -----
1. A new internal debug flag, which enables JNI weak reference check in the function that adds to cache.
2. A new Java functional test used to reproduce the problem.
3. The fix in JniWrapperCache, which ensures that if object was garbage collected then the handle is properly deleted and lost during override.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>