C++: generate synchronous version of function only when needed
When '@Async' annotation is used then C++ generator
behaves differently. In the past it was generating
two versions of the same function in such a case:
synchronous and asynchronous.

However, if the user only wants to generate the code
for C++ and Dart, then the synchronous version of a
function is not needed -- nobody is going to use it.

This change adjusts mustache templates and predicates
used for C++ generator in order to avoid generating
redundant code.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>