Java: add missing documentation comment to NativeBase
'propagateCleanupException' flag was added in the past
to NativeBase class. It allows controlling whether exceptions
related to cleanup of C++ objects should be propagated to
client application. This is an internal flag for developers.

Sadly, the comment was not added for this flag -- this could
trigger compilation warnings (or errors depending on the flags).

This change adds the missing comment.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>