JNI: check external field for used platform
JniTemplates class is used for two platforms: Java and Kotlin.
In the past it was used only for Java. Therefore, its mustache
files used 'external.java' explicitly.

This was the root cause of problems with generation of JNI layer
for Kotlin when external types with converters were used.

This commit adjusts mustache templates to use 'external' field
for the selected platform. This is achieved via new predicates
as well as usage of 'EvalHelper'.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>