JniTypeSignatureNameResolver: properly resolve fundamental kotlin types
Fundamental Kotlin types like 'Int' or 'String' are mapped to Java types
when compiled for JVM. The declaration of the usual function signatures
and types work as expected, because parsing of LIME files use 'TypeId'
for basic types.

However, the situation is much different if the user specifies the basic
type to act as an external type of a defined structure/class/enumeration.

Therefore, when resolving function signatures of external types,
which are set to fundamental Kotlin types we need to perform appropriate
mapping to avoid runtime issues related to missing functions.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>