functional tests: fix broken tests of 'Skip/EnableIf'
The functional tests define types that should be skipped
in 'functional/input/src' to trigger redefinition error
when 'Skip/EnableIf' works incorrectly.

Sadly, the input LIME files are incorrectly defined.
Even when 'Skip/EnableIf' is removed from the types,
which shall trigger compilation error they do not do
it.

The root cause of the problem is definition of the types
which should be skipped as nested types. The types defined
in 'functional/input/src' are not defined in nested types.

When the types in LIME files are defined as standalone types
and 'Skip/EnableIf' is removed, then compilation error is
correctly triggered.

Thus, this commit changes input LIME files to define the
interesting types as standalone ones.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>