functional tests: fix broken tests of 'Skip/EnableIf' (#1669)
The functional tests define types that should be skipped
in 'functional/input/src' to trigger redefinition error
when 'Skip/EnableIf' works incorrectly.

Sadly, the input LIME files are incorrectly defined.
Even when 'Skip/EnableIf' is removed from the types,
which shall trigger compilation error they do not do
it.

The root cause of the problem is definition of the types
which should be skipped as nested types. The types defined
in 'functional/input/src' are not defined as nested types.

When the types in LIME files are defined as standalone types
and 'Skip/EnableIf' is removed, then compilation error is
correctly triggered.

Thus, this commit changes input LIME files to define the
interesting types as standalone ones.

Moreover, this change adds missing files that should trigger compilation
error when 'Skip/EnableIf' works incorrectly in the generator for Kotlin.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>