Kotlin: test usage of collection fields (#1670)
This series brings new functional tests related to usage of collection fields for Kotlin generator:
 - ArraysTest.kt
 - MapsTest.kt
 - SetTypeTest.kt

Moreover it adds smoke tests for collections and enumerations. It also fixes the gap related to
lack of usage of EnumSet by Kotlin generator. The usage of the mentioned class is introduced
to ensure that the behavior is identical to Java generator.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>