Implement placeholders for comments (#1666)
----- Motivation -----
When creating documentation comments there are elements, which may be
used often. Sometimes the repetition leads to problems. Let's look at the following
example showing the code needed to properly word the interface for different outputs:

```
{@Java interface}{@Swift protocol}{@Dart abstract class}{@Cpp abstract class}
```

It is easy to forget adding wording for one of output platforms. Moreover, the mistake
may be repeated if this long syntax is copied to other places in the code.

----- Solution -----
The placeholders for comments are introduced to solve the described problem.
Instead of using the long syntax the new `{@Placeholder <PLACEHOLDER_NAME>}`
can be used.

The new configuration parameters, which allow setting placeholders file are available:
- CLI: `-docsplaceholderslist`
- CMake: `GLUECODIUM_DOCS_PLACEHOLDERS_LIST`

The file must follow .properties file format and must contain the known comment placeholders
as `placeholder_name="placeholder_value"`

Note: the placeholders respect the usual LIME markup -- it means that nested platforms tag can
be used, but they are not required.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>