functional tests: implement 'Dates' tests for Kotlin (#1677)
This change introduces functional tests for 'Dates'
test suite for Kotlin. The logic was inspired by the
functional tests suite for Java to ensure the same
level of support in Kotlin.

Moreover, handling of imports for type aliases was adjusted.
The usage of 'LimeTypeAlias' was not present in imports
resolver and imports collector. This change updates the code
to avoid problems with missing imports.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>