functional tests: implement 'method overloads' tests for Kotlin
This change introduces functional tests for 'method overloads' test
suite for Kotlin. The logic was inspired by the functional test suite
for Java to ensure the same level of support in Kotlin.

Note: two more test cases will be added when 'Errors' test suite is implemented
for Kotlin, because they use 'ThrowingConstructor' type, which is defined
in LIME files for errors.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>