Kotlin: support usage of external types (#1665)
This commit introduces a set of test cases for
usage of external types for Kotlin generated code.
It also fills the gaps in the existing implementation
to make the tests pass.

Two important notes:
 - fundamental numeric types of Kotlin are deliberately
   used with '?' suffix when defining the external types,
   because currently JNI templates assume that the used type
   is an object
- for fundamental types this assumption is true only if
  the nullable version is used; for instance 'Kotlin.Int'
  is mapped to trivial 'int' whether 'Kotlin.Int?' is mapped
  to 'java.lang.Integer'

The support for usage of non-nullable optional types as the
external types will be added in a follow-up PR, because it
requires changes in JNI mustache files

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>