Validate comment's placeholders during initialization (#1688)
The initial commit, which introduced usage of
placeholders checked if a placeholder exists
when getting the conent of comment. It was
problematic because the exact location of
a file, which contains invalid placeholder
was not known.

This commit adjusts that by moving the validation
to init block of LimeComment. This way LIME model
builder may throw ParseCancellationExcetion.

Before this change the error looked as follows when
invalid placeholder was used:

```
java.lang.IllegalArgumentException Invalid comment placeholder requested: Interface.
```

Now the stack contains the following information:
```
com.here.gluecodium.loader.LimeLoadingException: Syntax errors found, see log for details.
```

And the standard error contains the full path:
```
SEVERE: com.here.gluecodium.loader.LimeBasedLimeModelLoader loadFile:
    File /some/path/Comments.lime, line 20:0, Invalid comment placeholder requested: Interface.
```

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>