functional tests: implement 'errors' tests for Kotlin
This change introduces functional tests for 'errors' test suite for Kotlin.
The logic was inspired by the functional test suite for Java to ensure the
same level of support in Kotlin.

Note: this change enabled also compilation of 'Blob' LIME input
because certain types were required for errors test too.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>