JNI/Java: rename helper method to avoid using 'get' prefix
JNI and Java code used 'getSize()' helper method
for optimized lists. However, the usage of 'get'
prefix caused problems for Kotlin generated code.

The lists in Java provide 'size()' method for querying
the number of elements in the list. In Kotlin this is
represented as a property.

As the direct consequence for 'size' property Kotlin
expects 'getSize()' method -- we provide it in Kotlin
code via overriding the property. However, because of
the existence of the helper there is a method names
clash.

In order to solve the problem the helper method was
renamed from 'getSize()' to 'obtainSize()' to avoid
any overload problems.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>