Kotlin: support optimized lists (#1694)
This change introduces support for usage of optimized
lists. Two new mustache templates are added:
- KotlinAbstractNativeList.mustache
- KotlinLazyNativeList.mustache

Aside from it the code of KotlinGenerator is adjusted
to use OptimizedListCollector to firstly collect all
optimized lists and then feed the template engine
with them.

This commit adjusts also JNI/Java mustache templates
to rename 'getSize()' helper method to 'obtainSize()' in
order to avoid problems with 'size' property in Kotlin.
Kotlin searches for 'get()' when getting the property on JVM.
To avoid name clash protected helper was renamed.

This commit introduces functional tests and smoke tests for
the new functionality to confirm that the new code works as
expected.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>