Kotlin: correctly resolve visibility of elements (#1698)
This change brings handling of 'Internal' attribute
to properly resolve visibility modifier to 'internal'
Kotlin keyword.

Note: if internal type/function/property is defined
directly in an interface, then 'internal' Kotlin
keyword is not used, because the lanugage imposes
such limitation.

This change brings also smoke and functional tests
of the new functionality to showcase how it works.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>