lime-loader: properly parse 'AfterConstruction' attribute
Introduced parsing of a new attribute called
'AfterConstruction' which allows users to set
a custom function, which should be called when
the construction of an object finishes.

It must be a static function, which takes parameters
with names of params or constructor or special
'this' instance.

The parsing creates LimeFunction with proper
path, parameters and isConstructor flag in order
to use it in the mustache files as it was an
ordinary function.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>