Ensure that platform code called from constructor receives valid object (#1705)
----------------- Description of problem -----------------

Let's consider the following situation:
- we have a constructor, which takes an interface: 'constructor(someInterface: SomeInterface)'
- we want to call a method of the interface from the constructor with the newly created object:
   'someInterface.doSomething(this)'
- if the call is performed from C++ factory function implemented by the user of the generated code,
   then at the moment of call the platform object (Java, Swift, Dart) is not properly constructed and cached
- to avoid such problematic situations and breaking the cache/leaking objects the newly implemented
   'AfterConstruction()' attribute should be used

It was confirmed that in the case of Java weak reference leak happened in the described scenario.

----------------- Designed solution -----------------
In order to be able to use listeners in constructor the new attribute called AfterConstruction
has been implemented. It allows users to specify a 'hook' function that will be called for the
given object after it is properly created and cached. The user must implement the mentioned hook.

For more information please check docs/lime_attributes.md.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>