Use LimeLazyFunctionCall instead of LimeFunction
In order to have all the data about the called function that
is passed to 'AfterConstruction()' the new model type was
introduced: LimeLazyFunctionCall.

It wouldn't be possible to get the original parameter names
needed for Swift or the exception types.

The new type contains the path to called function and information about
parameters to know which one should be called with 'this'.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>