Dart: correctly handle callbacks executed from isolate thread outside of its context (#1712)
----- Description of the problem -----
Flutter 3.29 introduced changes in threading model.
The Dart code is executed in the main platform thread.
Before it was executed by the dedicated UI thread.

It is possible to schedule some work via platform-specific
API (e.g. Swift: DispatchQueue, Android: Looper) in the way
that it will be executed by the main platform thread, but it will
be outside of isolate's context. If such work tries to execute Dart's
callback from C++ we would see deadlock.

----- Implemented solution -----
The generated code uses NativeCallable.isolateLocal to obtain native
function pointers from Dart. Moreover, it stores isolate's thread id and
isolate's handle when constructing C++ callable. When we want to
execute the callback, then we check if we are in good thread. If we
are, but we are not in isolate context, then we manually enter it via Dart DL API.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>