OptionReader: implement parsing of LimeDocValidationRules
This change extends CLI of Gluecodium with new parameter
called -docsvalidationrules, which allows users to
specify a JSON file with list of validation rules.

Each validation rule has name, list of elements for which
it applies, regular expression that is checked and flag
that describes if complying with regex is required to
properly generate files.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>